// RookieEcommerce.AdminSite/src/api/models/ProductDetailsDtoPaginationList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProductDetailsDto } from './ProductDetailsDto';
import {
    ProductDetailsDtoFromJSON,
    ProductDetailsDtoFromJSONTyped,
    ProductDetailsDtoToJSON,
    ProductDetailsDtoToJSONTyped,
} from './ProductDetailsDto';

/**
 * 
 * @export
 * @interface ProductDetailsDtoPaginationList
 */
export interface ProductDetailsDtoPaginationList {
    /**
     * 
     * @type {Array<ProductDetailsDto>}
     * @memberof ProductDetailsDtoPaginationList
     */
    items?: Array<ProductDetailsDto> | null;
    /**
     * 
     * @type {number}
     * @memberof ProductDetailsDtoPaginationList
     */
    pageNumber?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductDetailsDtoPaginationList
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductDetailsDtoPaginationList
     */
    totalCount?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductDetailsDtoPaginationList
     */
    readonly totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ProductDetailsDtoPaginationList
     */
    readonly hasPreviousPage?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProductDetailsDtoPaginationList
     */
    readonly hasNextPage?: boolean;
}

/**
 * Check if a given object implements the ProductDetailsDtoPaginationList interface.
 */
export function instanceOfProductDetailsDtoPaginationList(value: object): value is ProductDetailsDtoPaginationList {
    return true;
}

export function ProductDetailsDtoPaginationListFromJSON(json: any): ProductDetailsDtoPaginationList {
    return ProductDetailsDtoPaginationListFromJSONTyped(json, false);
}

export function ProductDetailsDtoPaginationListFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductDetailsDtoPaginationList {
    if (json == null) {
        return json;
    }
    return {
        
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(ProductDetailsDtoFromJSON)),
        'pageNumber': json['pageNumber'] == null ? undefined : json['pageNumber'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
        'hasPreviousPage': json['hasPreviousPage'] == null ? undefined : json['hasPreviousPage'],
        'hasNextPage': json['hasNextPage'] == null ? undefined : json['hasNextPage'],
    };
}

export function ProductDetailsDtoPaginationListToJSON(json: any): ProductDetailsDtoPaginationList {
    return ProductDetailsDtoPaginationListToJSONTyped(json, false);
}

export function ProductDetailsDtoPaginationListToJSONTyped(value?: Omit<ProductDetailsDtoPaginationList, 'totalPages'|'hasPreviousPage'|'hasNextPage'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(ProductDetailsDtoToJSON)),
        'pageNumber': value['pageNumber'],
        'pageSize': value['pageSize'],
        'totalCount': value['totalCount'],
    };
}